requirements.txt
openai==0.26.5
flask==2.2.3
python-dotenv==0.21.1